manage files and renaming them
