+++ client/src/components/Navbar.js
@@ -49,7 +49,7 @@ const AppNavbar = () => {
               {Auth.loggedIn() ? (
                 <>
                   <Link
-                    to="/addQuiz"
+                    to="/profile"
                     className="ml-2 px-2 cursor-pointer text-xl  bg-green-800 py-1  rounded-lg text-green-100 no-underline hover:bg-green-100 hover:text-green-600"
                   >
                     Profile
