--- client/src/pages/AddQuiz.js
@@ -1,12 +1,0 @@
-import React from "react";
-import QuizSearchForm from "../components/Quiz/QuizSearchForm";
-
-const AddQuiz = () => {
-  return (
-    <div>
-      <QuizSearchForm />
-    </div>
-  );
-};
-
-export default AddQuiz;
